refactor(pricing): Move pricing and demo data to dedicated files for better organization
