+++ apps/website/src/components/EnhancedPricingTiers.tsx
@@ -3,30 +3,7 @@ import { motion, useReducedMotion, TargetAndTransition } from 'framer-motion';
 import { Check } from 'lucide-react';
 import { Link } from 'react-router-dom';
 import { Button } from './ui/button';
-
-interface PricingFeature {
-    name: string;
-    description?: string;
-    tiers: {
-        starter: boolean | string;
-        pro: boolean | string;
-        team: boolean | string;
-    };
-}
-
-interface PricingTier {
-    name: string;
-    description: string;
-    monthlyPrice: number;
-    annualPrice: number;
-    currency: string;
-    features: string[];
-    cta: {
-        text: string;
-        href: string;
-    };
-    popular?: boolean;
-}
+import { PricingTier, PricingFeature } from '../data/pricing/pricing-tiers';
 
 interface EnhancedPricingTiersProps {
     title?: string;
@@ -291,155 +268,3 @@ export function EnhancedPricingTiers({
         </section>
     );
 }
-
-// Example pricing tiers
-export const pricingTiers: PricingTier[] = [
-    {
-        name: 'Starter',
-        description:
-            'Perfect for individuals and small teams just getting started',
-        monthlyPrice: 29,
-        annualPrice: 24,
-        currency: 'USD',
-        features: [
-            'Up to 5 workflows',
-            'Basic integrations',
-            'Community support',
-            '1 user',
-            '1,000 AI credits/month',
-        ],
-        cta: {
-            text: 'Start free trial',
-            href: '/waitlist',
-        },
-    },
-    {
-        name: 'Pro',
-        description: 'For growing teams that need more power and flexibility',
-        monthlyPrice: 79,
-        annualPrice: 64,
-        currency: 'USD',
-        features: [
-            '10 workflows',
-            'Advanced integrations',
-            'Email support',
-            '1 user',
-            '5,000 AI credits/month',
-            'Workflow templates',
-            'Advanced analytics',
-        ],
-        cta: {
-            text: 'Start free trial',
-            href: '/waitlist',
-        },
-        popular: true,
-    },
-    {
-        name: 'Team',
-        description:
-            'For organizations that need advanced security and control',
-        monthlyPrice: 299,
-        annualPrice: 239,
-        currency: 'USD',
-        features: [
-            'Unlimited workflows',
-            'Custom integrations',
-            'Dedicated support',
-            '5 users',
-            '20,000 AI credits/month',
-            'SSO & SAML (Coming soon)',
-            'Audit logs',
-            'Custom training',
-        ],
-        cta: {
-            text: 'Contact sales',
-            href: '/contact',
-        },
-    },
-];
-
-// Example feature comparison
-export const featureComparison: PricingFeature[] = [
-    {
-        name: 'Workflows',
-        description: 'Create automated sequences of actions',
-        tiers: {
-            starter: 'Up to 5',
-            pro: '10',
-            team: 'Unlimited',
-        },
-    },
-    {
-        name: 'Users',
-        tiers: {
-            starter: '1',
-            pro: '1',
-            team: '5',
-        },
-    },
-    {
-        name: 'AI Credits',
-        description: 'Used for AI-powered features',
-        tiers: {
-            starter: '1,000/month',
-            pro: '5,000/month',
-            team: '20,000/month',
-        },
-    },
-    {
-        name: 'Integrations',
-        tiers: {
-            starter: 'Basic',
-            pro: 'Advanced',
-            team: 'Custom',
-        },
-    },
-    {
-        name: 'Support',
-        tiers: {
-            starter: 'Community',
-            pro: 'Email',
-            team: 'Dedicated',
-        },
-    },
-    {
-        name: 'SSO & SAML',
-        tiers: {
-            starter: false,
-            pro: false,
-            team: 'Coming soon',
-        },
-    },
-    {
-        name: 'Audit Logs',
-        tiers: {
-            starter: false,
-            pro: false,
-            team: true,
-        },
-    },
-    {
-        name: 'Custom Training',
-        tiers: {
-            starter: false,
-            pro: false,
-            team: true,
-        },
-    },
-    {
-        name: 'Workflow Templates',
-        tiers: {
-            starter: false,
-            pro: true,
-            team: true,
-        },
-    },
-    {
-        name: 'Advanced Analytics',
-        tiers: {
-            starter: false,
-            pro: true,
-            team: true,
-        },
-    },
-];
