+++ packages/ui/src/components/badge.tsx
@@ -1,22 +1,7 @@
 import * as React from 'react';
 import { type VariantProps } from 'class-variance-authority';
-import { cva } from 'class-variance-authority';
 import { cn } from '@vernisai/utils';
-
-const badgeVariants = cva(
-    'inline-flex items-center rounded-full px-2.5 py-0.5 text-xs font-medium',
-    {
-        variants: {
-            variant: {
-                default: 'bg-gray-100',
-                outline: 'border border-gray-200',
-            },
-        },
-        defaultVariants: {
-            variant: 'default',
-        },
-    },
-);
+import { badgeVariants } from './badge-variants';
 
 export interface BadgeProps
     extends React.HTMLAttributes<HTMLSpanElement>,
@@ -35,4 +20,4 @@ const Badge = React.forwardRef<HTMLSpanElement, BadgeProps>(
 );
 Badge.displayName = 'Badge';
 
-export { Badge, badgeVariants };
+export { Badge };
